+++ game/cards/dm05/dragonoid.go
@@ -1,0 +1,22 @@
+package dm05
+
+import (
+	"duel-masters/game/civ"
+	"duel-masters/game/family"
+	"duel-masters/game/fx"
+	"duel-masters/game/match"
+)
+
+// BombatGeneralOfSpeed ...
+func BombatGeneralOfSpeed(c *match.Card) {
+
+	c.Name = "Bombat, General of Speed"
+	c.Power = 3000
+	c.Civ = civ.Fire
+	c.Family = family.Dragonoid
+	c.ManaCost = 5
+	c.ManaRequirement = []string{civ.Fire}
+
+	c.Use(fx.Creature, fx.SpeedAttacker)
+
+}